+++ devel/CHANGELOG.md
@@ -10,6 +10,11 @@
 - [trvl]: trivial modifications;
 - [warn]: fixing compiler warning messages;
 
+## 2022-03-11
+- JNIQ: adding missing functions,
+- N4JRE, SCLJRE: introducing `jCore`,
+- TOL: `qDELETE(pointer)` sets now pointer to `NULL`,
+
 ## 2022-03-09
 - BCH:
   - typo,